move tests; remove unnecessary using statements
